% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_intercept_and_centering.R
\name{add_intercept}
\alias{add_intercept}
\title{Add Random Intercepts to Synthetic Data}
\usage{
add_intercept(tmp, method = "uniform")
}
\arguments{
\item{tmp}{List. Output from synthetic generator containing raw and smooth matrices.}

\item{method}{Character. Distribution method for random intercepts ('uniform', 'normal', 'uniform01', 'normal_adapt').}
}
\value{
A list containing modified raw and smooth matrices.
}
\description{
Add Random Intercepts to Synthetic Data
}
