% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_main.R
\name{C.calculation}
\alias{C.calculation}
\title{C-statistic Calculation for GTFS}
\usage{
C.calculation(scheme = "GTFS", xfd, N, H, nbasis, nharm, v.prop, q.star = 0.1)
}
\arguments{
\item{scheme}{GTFS variants selection ('GTFS', 'GTFS(P)', 'scheme1', 'scheme2')}

\item{xfd}{input matrix converted into fd object}

\item{N}{sample size}

\item{H}{clean subset}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{q.star}{Gumbel cutoff value described in the manuscript ('default' gives 2/sqrt(N) as described in the manuscript)}
}
\value{
A list containing:
\item{score.H}{Numeric vector of the calculated robust outlyingness scores for all N observations.}
\item{fpca}{The raw Functional Principal Component Analysis (FPCA) object returned from \code{pca.fd}.}
\item{d}{Integer. The number of principal components explaining the pre-specified proportion of total variation.}
\item{d.select}{Integer. The number of filtered/selected eigenfunctions after thresholding with the Gumbel cutoff.}
}
\description{
C-statistic Calculation for GTFS
}
