% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheme2_calculation.R
\name{D.calculation}
\alias{D.calculation}
\title{Score calculation}
\usage{
D.calculation(xfd, H, N, nbasis, nharm, adjust = T, re.adjust = F, v.prop)
}
\arguments{
\item{xfd}{input matrix converted into fd object}

\item{H}{subset index}

\item{N}{sample size}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{adjust}{eigenvalue adjustment described in the manuscript}

\item{re.adjust}{additional adjustment option}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}
}
\value{
A list containing:
\item{delta.H}{Numeric vector of the calculated robust outlyingness scores for all N observations.}
\item{r}{Integer. The number of principal components explaining the pre-specified proportion of total variation.}
\item{fpca}{The raw Functional Principal Component Analysis (FPCA) object returned from \code{pca.fd}.}
}
\description{
Score calculation
}
