% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheme1_calculation.R
\name{delta.calculation}
\alias{delta.calculation}
\title{Score calculation}
\usage{
delta.calculation(xfd, H, N, nbasis, nharm, v.prop)
}
\arguments{
\item{xfd}{input matrix converted into fd object}

\item{H}{subset index}

\item{N}{sample size}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}
}
\value{
A list containing:
\item{delta.H}{Numeric vector of the calculated robust outlyingness scores for all N observations.}
\item{r}{Integer. The number of principal components explaining the pre-specified proportion of total variation.}
\item{fpca}{The raw Functional Principal Component Analysis (FPCA) object returned from \code{pca.fd}.}
\item{e.c}{Centered functions}
}
\description{
Score calculation
}
