% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheme1_calculation.R
\name{find_cutoff}
\alias{find_cutoff}
\title{Finding a threshold incorporating the chi-square convolution}
\usage{
find_cutoff(r, ec.pc, alpha, delta.H, H, adjust = T, re.adjust = F)
}
\arguments{
\item{r}{number of incorporated eigenfunctions}

\item{ec.pc}{FPCA result obtained from centered data}

\item{alpha}{expected false positive rate for threshold}

\item{delta.H}{calculated scores}

\item{H}{subset indice}

\item{adjust}{eigenvalue adjustment described in the manuscript}

\item{re.adjust}{additional adjustment option}
}
\value{
A list containing:
\item{cutoff}{Numeric. The theoretical outlier detection threshold determined by the Chi-Square distribution.}
\item{density}{Numeric vector. The density estimate for the outlyingness statistics used for empirical distribution tracking.}
}
\description{
Finding a threshold incorporating the chi-square convolution
}
