% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_main.R
\name{GTFS_outlier}
\alias{GTFS_outlier}
\title{GTFS calculation function}
\usage{
GTFS_outlier(
  scheme = "GTFS",
  x,
  t,
  basis = "bspline",
  nbasis = 31,
  nharm = 20,
  v.prop = 0.9,
  alpha = 0.05,
  q.star = "default",
  n.init = 5,
  iter.max = 10,
  refined = T
)
}
\arguments{
\item{scheme}{GTFS variants selection ('GTFS', 'GTFS(P)', 'scheme1', 'scheme2')}

\item{x}{N x p functional data}

\item{t}{sampling time grid points}

\item{basis}{smoothing basis ('bspline' or 'fourier')}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{alpha}{expected false positive rate}

\item{q.star}{Gumbel cutoff value described in the manuscript ('default' gives 2/sqrt(N) as described in the manuscript)}

\item{n.init}{number of random multi-starts}

\item{iter.max}{maximum number of iteration}

\item{refined}{whether to conduct one-step refinement described in manuscript}
}
\value{
A list containing:
\item{H}{Integer vector. The final optimized clean subset indices of size h used for robust estimation.}
\item{score.H}{Numeric vector. The calculated robust functional outlyingness scores for all N observations.}
\item{cutoff}{Numeric. The theoretical outlier detection threshold determined by the Chi-Square distribution.}
\item{outlier}{Integer vector. The final detected outlier indices whose scores exceed the cutoff threshold.}
\item{density}{Numeric vector. The density estimate for the outlyingness statistics used for empirical distribution tracking.}
\item{d}{Integer. The original number of estimated eigenfunctions explaining the pre-specified variation proportion.}
\item{d.select}{Integer. The reduced number of filtered eigenfunctions actively used after Gumbel thresholding.}
\item{step.iter}{Integer. The number of concentration iterations taken until the subset convergence criteria was met.}
}
\description{
GTFS calculation function
}
