% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_main.R
\name{gtfs_preprocess}
\alias{gtfs_preprocess}
\title{Preprocessing for GTFS including integral centering}
\usage{
gtfs_preprocess(x, t, basis = "bspline", nbasis = 20)
}
\arguments{
\item{x}{N x p functional data}

\item{t}{sampling time grid points}

\item{basis}{smoothing basis ('bspline' or 'fourier')}

\item{nbasis}{number of basis functions for smoothing}
}
\value{
A list containing:
\item{xfd}{x converted into fd object}
\item{x.ic}{Matrix. xfd curves evaluated on measurement points t}
}
\description{
Preprocessing for GTFS including integral centering
}
