% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_intercept_and_centering.R
\name{integral_centering}
\alias{integral_centering}
\title{Perform quick Integral Centering on Functional Data (non-smoothed ver.)}
\usage{
integral_centering(X)
}
\arguments{
\item{X}{Matrix. N x p matrix of functional data.}
}
\value{
Centered functional matrix.
}
\description{
Perform quick Integral Centering on Functional Data (non-smoothed ver.)
}
