% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_main.R
\name{Iter.step.C}
\alias{Iter.step.C}
\title{Iteration step function}
\usage{
Iter.step.C(
  scheme = "GTFS",
  xfd,
  N,
  H,
  h,
  nbasis,
  nharm = 10,
  v.prop = 0.9,
  q.star = 0.1,
  init.pca,
  init.coeffs = NULL
)
}
\arguments{
\item{scheme}{GTFS variants selection ('GTFS', 'GTFS(P)', 'scheme1', 'scheme2')}

\item{xfd}{input matrix converted into fd object}

\item{N}{sample size}

\item{H}{clean subset}

\item{h}{prespecified clean subset size}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{q.star}{Gumbel cutoff value described in the manuscript ('default' gives 2/sqrt(N) as described in the manuscript)}

\item{init.pca}{initial FPCA estimates}

\item{init.coeffs}{initial selection coefficients}
}
\value{
A list containing:
\item{H1}{Integer vector containing the updated subset indices of size h with the smallest sum of delta.}
\item{score.H}{Numeric vector of the updated outlyingness scores for all N observations.}
\item{coeffs}{Numeric vector of the updated weights/coefficients applied to each eigenfunction score.}
}
\description{
Iteration step function
}
