% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheme1_calculation.R
\name{Iter.step.delta}
\alias{Iter.step.delta}
\title{Iteration step function}
\usage{
Iter.step.delta(xfd, N, H, h, nbasis, v.prop = 0.9, nharm = 10, init.pca)
}
\arguments{
\item{xfd}{input matrix converted into fd object}

\item{N}{sample size}

\item{H}{subset index}

\item{h}{subset size}

\item{nbasis}{number of basis functions for smoothing}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{init.pca}{initial FPCA estimates}
}
\value{
A list containing:
\item{H1}{Integer vector containing the updated subset indices of size h with the smallest sum of delta.}
\item{sum.delta.H}{Numeric. The sum of outlyingness scores within the updated subset H1.}
\item{delta.H}{Numeric vector of the updated outlyingness scores for all N observations.}
}
\description{
Iteration step function
}
