% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFS_calculation.R
\name{LTFS}
\alias{LTFS}
\title{Least Trimmed Functional Score (LTFS) Algorithm}
\usage{
LTFS(mdp, m1, m2, C, N, h)
}
\arguments{
\item{mdp}{Integer vector. Robust initial subset indices (e.g., from MDP algorithm).}

\item{m1}{Integer. Number of random initial subsets of size 2.}

\item{m2}{Integer. Number of best candidate subsets to keep for the final C-step.}

\item{C}{Matrix. The \code{N x p} functional/high-dimensional data matrix.}

\item{N}{Integer. Total sample size.}

\item{h}{Integer. The final target clean subset size (typically \code{[N/2] + 1}).}
}
\value{
A list containing:
\text{Hltfs} {The final optimized LTFS subset indices.}
\text{DH} {The robust functional outlyingness score for all N observations.}
\text{dltfs} {The number of eigenfunctions explaining up to 90\\% of variation.}
}
\description{
Least Trimmed Functional Score (LTFS) Algorithm
}
\references{
Ren, H., Chen, N., & Zou, C. (2017). Projection-based outlier detection
in functional data. \emph{Biometrika}, 104(2), 411-423.
}
