% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFS_calculation.R
\name{LTFS_benchmark}
\alias{LTFS_benchmark}
\title{Least Trimmed Functional Score (LTFS) Benchmark Toolkit}
\description{
This file contains the official benchmark implementation of the LTFS algorithm
for functional outlier detection, as proposed by Ren, et. al. (2017) Biometrika.
}
\author{
Ren, Haojie, Nan Chen, and Changliang Zou
}
