% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_mod}
\alias{plot_mod}
\title{Plot Generated Curves and Spot Outliers}
\usage{
plot_mod(Y, main = "", n = 200, c = 0.025)
}
\arguments{
\item{Y}{Matrix. Functional data matrix where rows represent curves.}

\item{main}{Character. Title of the plot.}

\item{n}{Integer. Number of observations.}

\item{c}{Numeric. Contamination ratio.}
}
\description{
Plot Generated Curves and Spot Outliers
}
