% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFS_calculation.R
\name{Refined}
\alias{Refined}
\title{One-Step Refined Estimation for LTFS Algorithm}
\usage{
Refined(N, od, allC)
}
\arguments{
\item{N}{Integer. Total sample size.}

\item{od}{Integer vector. Indices of preliminary outliers detected by LTFS.}

\item{allC}{Matrix. The complete \code{N x p} functional data matrix.}
}
\value{
A list containing the refined distance statistics and updated dimension d.
}
\description{
One-Step Refined Estimation for LTFS Algorithm
}
