% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheme1_calculation.R, R/Scheme2_calculation.R
\name{scheme1_core}
\alias{scheme1_core}
\title{GTFS variant (Scheme1)}
\description{
This file contains the implementation of the GTFS variant 'Scheme1'
for functional shape outlier detection.

This file contains the implementation of the GTFS variant 'Scheme2'
for functional shape outlier detection.
}
\author{
Hyungjun Lim
}
