% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme1_tools.R
\name{scheme1_cutoff}
\alias{scheme1_cutoff}
\title{Cutoff value calculation function}
\usage{
scheme1_cutoff(d, e.c.pc, alpha, score.H, adjust = T)
}
\arguments{
\item{d}{number of eigenfunction}

\item{e.c.pc}{fpca result}

\item{alpha}{expected false positive rate}

\item{score.H}{GTFS score}

\item{adjust}{eigenvalue adjustment described in the manuscript}
}
\value{
A list containing:
\item{cutoff}{Numeric. The theoretical outlier detection threshold determined by the Chi-Square distribution.}
\item{density}{Numeric vector. The density estimate for the outlyingness statistics used for empirical distribution tracking.}
}
\description{
Cutoff value calculation function

Scheme 1 cutoff calculation
}
\author{
Hyungjun Lim
}
