% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_generators.R
\name{simul_31}
\alias{simul_31}
\title{Generate Synthetic altered Slope Outliers (Model 3-1)}
\usage{
simul_31(n, p, t, c)
}
\arguments{
\item{n}{Integer. Total number of samples.}

\item{p}{Integer. Number of time points (dimensions).}

\item{t}{Numeric vector. Time evaluation points.}

\item{c}{Numeric. Contamination rate (between 0 and 1).}
}
\value{
A list containing two matrices: \code{raw} (noise-free) and \code{smooth} (Matern-smoothed).
}
\description{
This function generates altered version of the slope outliers proposed by Dai, Wenlin , et al. (2020).
Non-outliers are redesigned to share a common mean function.
It returns both noise-free raw data and smoothed data using a Matern covariance structure.
}
\references{
Dai, Wenlin, et al. (2020) "Functional outlier detection and taxonomy by sequential transformations."
\emph{Computational Statistics & Data Analysis} 149, 106960.
}
