% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_calculation.R
\name{C.calculation}
\alias{C.calculation}
\title{C-statistic Calculation for GTFS}
\usage{
C.calculation(
  xfd,
  H,
  N,
  nbasis,
  nharm,
  adjust = T,
  re.adjust = F,
  v.prop,
  alpha.p = 0.2,
  gumbel.cutoff = 0.01
)
}
\arguments{
\item{xfd}{input matrix converted into fd object}

\item{H}{subset indice}

\item{N}{sample size}

\item{nbasis}{number of basis functions for smoothing}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{adjust}{eigenvalue adjustment described in the manuscript}

\item{re.adjust}{additional adjustment option}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{alpha.p}{Gumbel quantile for selection cutoff}

\item{gumbel.cutoff}{sample size dependent gumbel cutoff value described in the manuscript}
}
\value{
A list containing:
\item{delta.H}{Numeric vector of the calculated robust outlyingness scores for all N observations.}
\item{d}{Integer. The number of principal components explaining the pre-specified proportion of total variation.}
\item{fpca}{The raw Functional Principal Component Analysis (FPCA) object returned from \code{pca.fd}.}
\item{dd}{Integer. The number of filtered/selected eigenfunctions after thresholding with the Gumbel cutoff.}
}
\description{
C-statistic Calculation for GTFS
}
