% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_calculation.R
\name{GTFS_core}
\alias{GTFS_core}
\title{Generalized Trimmed Functional Score (GTFS)}
\description{
This file contains the implementation of the GTFS algorithm
for functional shape outlier detection.
}
\author{
Hyungjun Lim
}
