% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFS_calculation.R
\name{Iter.step.C}
\alias{Iter.step.C}
\title{Iteration step function}
\usage{
Iter.step.C(
  xfd,
  N,
  H,
  h,
  nbasis,
  v.prop = 0.9,
  nharm = 10,
  init.pca,
  alpha.p = 0.2,
  gumbel.cutoff = 0.01,
  init.coeffs = NULL,
  coeffs.type = "practical"
)
}
\arguments{
\item{xfd}{input matrix converted into fd object}

\item{N}{sample size}

\item{H}{subset index}

\item{h}{subset size}

\item{nbasis}{number of basis functions for smoothing}

\item{v.prop}{proportion of total variation expected to be explained by d eigenfunctions}

\item{nharm}{number of eigenfunctions to be estimated in FPCA}

\item{init.pca}{initial FPCA estimates}

\item{alpha.p}{Gumbel quantile for selection cutoff}

\item{gumbel.cutoff}{sample size dependent gumbel cutoff value described in the manuscript}

\item{init.coeffs}{initial selection coefficients}

\item{coeffs.type}{'practical' for GTFS(P) and 'algorithm' for GTFS}
}
\value{
A list containing:
\item{H1}{Integer vector containing the updated subset indices of size h with the smallest sum of delta.}
\item{sum.delta.H}{Numeric. The sum of outlyingness scores within the updated subset H1.}
\item{delta.H}{Numeric vector of the updated outlyingness scores for all N observations.}
\item{coeffs}{Numeric vector of the updated weights/coefficients applied to each eigenfunction score.}
}
\description{
Iteration step function
}
